+++ NEWS.md
@@ -1,3 +1,7 @@
+# DrBristol (Version: 0.0.1.1000)
+
+* Adding ability to specify size of rival urn to enable omnibus tests of working versus any rival.
+
 # DrBristol (Version: 0.0.1.0000)
 
 * Alpha version with full functionality as of Lopez and Bowers (2025)